Refactor actions to extract schemas, add addConvos action
